requirements-lock.txt
alembic==1.16.4
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
dnspython==2.7.0
email-validator==2.0.0.post2
Flask==2.3.3
gunicorn==21.2.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Babel==2.0.0
graphviz==0.21
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pipdeptree==2.28.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==2.3.7
WTForms==3.2.1

